PythonTgClient/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
telethon==1.34.0